requirements.txt
azure-functions
python-telegram-bot==8.1.1